+++ src/gtirb-builder/GtirbBuilder.cpp
@@ -20,7 +20,6 @@
 //  endorsement should be inferred.
 //
 //===----------------------------------------------------------------------===//
-
 #include "./GtirbBuilder.h"
 
 #include "./ElfReader.h"