+++ src/gtirb-builder/PeReader.cpp
@@ -142,7 +142,8 @@ void PeReader::addEntryBlock()
 void PeReader::addAuxData()
 {
     // Add `binaryType' aux data table.
-    std::vector<std::string> BinaryType = {"EXEC"};
+    bool DLL = Pe->header().has_characteristic(LIEF::PE::HEADER_CHARACTERISTICS::IMAGE_FILE_DLL);
+    std::vector<std::string> BinaryType = {"EXEC", DLL ? "DLL" : "EXE"};
     Module->addAuxData<gtirb::schema::BinaryType>(std::move(BinaryType));
 
     // Add `libraries' aux data table.