+++ src/gtirb-decoder/core/ModuleLoader.cpp
@@ -48,7 +48,7 @@ void ModuleLoader(const gtirb::Module& Module, DatalogProgram& Program)
     // For now, use ISA info to infer endianness.
     // TODO: Get the endianness info from gtirb.
     std::string Endianness;
-    if(BinaryIsa == "ARM" || BinaryIsa == "X64")
+    if(BinaryIsa == "ARM" || BinaryIsa == "X64" || BinaryIsa == "X86")
     {
         Endianness = "LE";
     }
@@ -109,6 +109,8 @@ const char* binaryISA(gtirb::ISA Arch)
 {
     switch(Arch)
     {
+        case gtirb::ISA::IA32:
+            return "X86";
         case gtirb::ISA::X64:
             return "X64";
         case gtirb::ISA::ARM: