+++ src/gtirb-decoder/target/PeX64Loader.h
@@ -32,7 +32,7 @@
 
 CompositeLoader PeX64Loader()
 {
-    CompositeLoader Loader("souffle_disasm_x64");
+    CompositeLoader Loader("souffle_disasm_x86_64");
     Loader.add(ModuleLoader);
     Loader.add(SectionLoader);
     Loader.add<X64Loader>();