+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+
+* Add preliminary x86-32 support.
+
 # 1.3.0
 
 * Populate GTIRB symbolic expression attributes.