+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+
+* Support ELF x64 static binaries
+* Add preliminary x86-32 support.
+
 # 1.3.0
 
 * Populate GTIRB symbolic expression attributes.