+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+* Support ELF x64 static binaries
 * Add preliminary x86-32 support.
 
 # 1.3.0